+++ lib/jsonparser.js
@@ -8,3 +8,19 @@ module.exports.parse = function (str) {
 
   return result
 }
+
+'use strict';
+const LosslessJSON = require('lossless-json')
+const Decimal = require('decimal.js')
+
+// convert LosslessNumber to Decimal
+function reviver (key, value) {
+  if (value && value.isLosslessNumber) {
+    return new Decimal(value.toString())
+  }
+  return value
+}
+
+module.exports.losslessParse = function(str) {
+	return LosslessJSON.parse(str, reviver)
+}
