+++ lib/transports/elasticsearch.js
@@ -4,18 +4,6 @@ var parseBaseURL = require('../parse-base-url')
 var aws4signer = require('../aws4signer')
 var async = require('async')
 
-'use strict';
-const LosslessJSON = require('lossless-json')
-const Decimal = require('decimal.js')
-
-// convert LosslessNumber to Decimal
-function reviver (key, value) {
-  if (value && value.isLosslessNumber) {
-    return new Decimal(value.toString())
-  }
-  return value
-}
-
 var elasticsearch = function (parent, url, options) {
   this.base = parseBaseURL(url, options.index)
   this.parent = parent
@@ -171,8 +159,8 @@ elasticsearch.prototype.getData = function (limit, offset, callback) {
         callback(err, [])
         return
       }
-      
-      var body = LosslessJSON.parse(response.body, reviver)
+
+      var body = jsonParser.losslessParse(response.body)
       self.lastScrollId = body._scroll_id
 
       if (self.lastScrollId === undefined) {
